+++ raiden-dapp/tests/unit/components/navigation/transfer-steps.spec.ts
@@ -93,6 +93,10 @@ describe('TransferSteps.vue', () => {
   }
 
   beforeAll(() => {
+    store.commit(
+      'userDepositTokenAddress',
+      '0x3a989D97388a39A0B5796306C615d10B7416bE77'
+    );
     store.commit('updateTokens', {
       '0x3a989D97388a39A0B5796306C615d10B7416bE77': {
         address: '0x3a989D97388a39A0B5796306C615d10B7416bE77',