+++ raiden-dapp/CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ### Added
 
+- [#1421] Dialog to withdraw from the user deposit.
 - [#168] Notification panel
-
 
 ### Fixed
 
@@ -18,6 +18,7 @@
 [#1610]: https://github.com/raiden-network/light-client/issues/1610
 [#1540]: https://github.com/raiden-network/light-client/issues/1540
 [#1579]: https://github.com/raiden-network/light-client/issues/1579
+[#1421]: https://github.com/raiden-network/light-client/issues/1421
 [#168]: https://github.com/raiden-network/light-client/issues/168
 
 ## [0.9.0] - 2020-05-28